refactor(plugin-twitter): split readTweet into modular structure
Split the monolithic readTweet.ts file into focused modules:
- types.ts: All TypeScript type definitions
- utils.ts: Utility functions for tweet processing
- clientUtils.ts: Twitter client management
- readTweet.ts: Core handler logic (45% reduction in size)

Benefits:
- Better code organization by concern
- Improved maintainability and readability
- Enhanced reusability of utility functions
- All 145 tests passing
- 96.57% statement coverage maintained
